tvOS
remove codes prohibited in tvOS.

+++ uikit_ext/page/S9CoverFlowView.m
@@ -56,7 +56,7 @@
 			[subview rollWithRotation:-rotate]; // rotating around axis Y
 			
 			// 2. scale
-			scale = expf(-fabsf(distance)) * SCALE_RANGE + SCALE_MIN;
+			scale = expf(-fabs(distance)) * SCALE_RANGE + SCALE_MIN;
 			[subview scaleWithScale:scale];
 			
 			// 3. position
@@ -78,7 +78,7 @@
 			[subview pitchWithRotation:rotate]; // rotating around axis X
 			
 			// 2. scale
-			scale = expf(-fabsf(distance)) * SCALE_RANGE + SCALE_MIN;
+			scale = expf(-fabs(distance)) * SCALE_RANGE + SCALE_MIN;
 			[subview scaleWithScale:scale];
 			
 			// 3. position
